Example dataset of % days of stress

For refugia definition 5: fertilization success

Usage

percentdays

Format

A data frame with 6 columns, and 697524 rows:

cellID

cellID corresponding to the cell index position of the raster of California cropped to 100m isobath

site

Longitude and latitude of the cell center

model

Earth System Model. Either "gfdltv", "hadtv", "ipsltv" or "ens"

refugiadays

Number of days classified as refugia for that year

percent

As above, but classed as a percentage value

year

Year of interest

Examples

head(percentdays)
#>   cellID            site refugiadays percent year  model
#> 1  63938 -120.702_34.576         365     100 1990 gfdltv
#> 2  63939 -120.669_34.576         365     100 1990 gfdltv
#> 3  64225 -120.669_34.542         365     100 1990 gfdltv
#> 4  64226 -120.635_34.542         365     100 1990 gfdltv
#> 5  64227 -120.602_34.542         365     100 1990 gfdltv
#> 6  64228 -120.569_34.542         365     100 1990 gfdltv
str(percentdays)
#> 'data.frame':	697524 obs. of  6 variables:
#>  $ cellID     : num  63938 63939 64225 64226 64227 ...
#>  $ site       : chr  "-120.702_34.576" "-120.669_34.576" "-120.669_34.542" "-120.635_34.542" ...
#>  $ refugiadays: num  365 365 365 365 365 365 365 365 365 365 ...
#>  $ percent    : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ year       : num  1990 1990 1990 1990 1990 1990 1990 1990 1990 1990 ...
#>  $ model      : chr  "gfdltv" "gfdltv" "gfdltv" "gfdltv" ...